+++ Maccy/Extensions/Defaults.Keys+Names.swift
@@ -47,6 +47,7 @@ extension Defaults.Keys {
   static let keepPreviewOpen = Key<Bool>("keepPreviewOpen", default: true)
   static let removeFormattingByDefault = Key<Bool>("removeFormattingByDefault", default: false)
   static let searchMode = Key<Search.Mode>("searchMode", default: .exact)
+  static let selectionMode = Key<SelectionMode>("selectionMode", default: .hover)
   static let showFooter = Key<Bool>("showFooter", default: true)
   static let showInStatusBar = Key<Bool>("showInStatusBar", default: true)
   static let showRecentCopyInMenuBar = Key<Bool>("showRecentCopyInMenuBar", default: false)